231: added cobbcounty modules to tabrasa composer
Composer alphabetized the modules. If you prefer me to update the composer.json to prevent the automatic alphabetizing, let me know.

I used drupal/config_ignore:2.x-dev because that is the currently recommended major version, and what's being used on cobbcounty, but they are working on a 3.x-dev version which we could test instead if you prefer.

https://git.ifsight.net/proj/requests/if-initiatives/issues/231